pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Policy-driven input wrapper with taint tracking and sink patches"
readme = "README.md"
authors = [
    { name = "sach-12", email = "sbalasubramanyam@ucdavis.edu" },
    { name = "Sarika-dinesh", email = "sarikadinesh@ucdavis.edu"},
    { name = "shreyaslshah", email = "slshah@ucdavis.edu"}
]
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "jsonschema>=4.22.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "pydantic>=2.9.2",
    "requests>=2.31.0",
    "jinja2>=3.1.2",
]

[project.scripts]
sentinel = "sentinel:main"

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "httpx>=0.27.0",
]
